move workspace billing to the billing info router
