+++ fixbackend/billing_information/schemas.py
@@ -16,8 +16,13 @@ from datetime import datetime
 
 from pydantic import BaseModel, Field
 
+from typing import Optional
+
 from fixbackend.ids import BillingId, SecurityTier, SubscriptionId, WorkspaceId
 from fixbackend.subscription.models import BillingEntry
+from enum import Enum
+
+from fixbackend.workspaces.models import Workspace
 
 
 class BillingEntryRead(BaseModel):
@@ -39,4 +44,46 @@ class BillingEntryRead(BaseModel):
             period_start=billing_entry.period_start,
             period_end=billing_entry.period_end,
             nr_of_accounts_charged=billing_entry.nr_of_accounts_charged,
+        )
+
+
+class PaymentMethod(str, Enum):
+    AWS = "aws_marketplace"
+
+
+class SecurityTierJson(str, Enum):
+    Free = "free"
+    Foundational = "foundational"
+    HighSecurity = "high_security"
+
+
+class WorkspaceBillingSettings(BaseModel):
+    payment_method: Optional[PaymentMethod] = Field(description="The payment method used for this workspace")
+    security_tier: SecurityTierJson = Field(description="The security tier of this workspace")
+
+    model_config = {
+        "json_schema_extra": {
+            "examples": [
+                {
+                    "payment_method": "aws_marketplace",
+                    "security_tier": "free",
+                }
+            ]
+        }
+    }
+
+    @staticmethod
+    def from_model(workspace: Workspace) -> "WorkspaceBillingSettings":
+        def tier(workspace: Workspace) -> SecurityTierJson:
+            match workspace.security_tier:
+                case SecurityTier.Free:
+                    return SecurityTierJson.Free
+                case SecurityTier.Foundational:
+                    return SecurityTierJson.Foundational
+                case SecurityTier.HighSecurity:
+                    return SecurityTierJson.HighSecurity
+
+        return WorkspaceBillingSettings(
+            payment_method=PaymentMethod.AWS,
+            security_tier=tier(workspace),
         )
