+++ fixbackend/workspaces/router.py
@@ -22,24 +22,20 @@ from fixbackend.auth.depedencies import AuthenticatedUser
 from fixbackend.auth.models import User
 from fixbackend.auth.user_repository import UserRepositoryDependency
 from fixbackend.config import ConfigDependency
-from fixbackend.ids import InvitationId, SubscriptionId, UserId, WorkspaceId, SecurityTier
-from fixbackend.subscription.subscription_repository import SubscriptionRepositoryDependency
+from fixbackend.ids import InvitationId, UserId, WorkspaceId
 from fixbackend.workspaces.invitation_service import InvitationServiceDependency
 from fixbackend.workspaces.repository import WorkspaceRepositoryDependency
 from fixbackend.workspaces.dependencies import UserWorkspaceDependency
 from fixbackend.workspaces.schemas import (
     ExternalIdRead,
     UserInvite,
-    WorkspaceBilling,
     WorkspaceCreate,
     WorkspaceInviteRead,
     WorkspaceRead,
     WorkspaceSettingsRead,
     WorkspaceSettingsUpdate,
     WorkspaceUserRead,
-    SecurityTierJson,
 )
-from fixbackend.errors import ResourceNotFound
 import asyncio
 
 
@@ -179,47 +175,6 @@ def workspaces_router() -> APIRouter:
         invitation = await invitation_service.accept_invitation(token)
         url = request.base_url.replace_query_params(message="invitation-accepted", workspace_id=invitation.workspace_id)
         return RedirectResponse(url)
-
-    @router.get("/{workspace_id}/billing")
-    async def get_billing(workspace: UserWorkspaceDependency) -> WorkspaceBilling:
-        """Get a workspace billing."""
-        return WorkspaceBilling.from_model(workspace)
-
-    @router.put("/{workspace_id}/billing")
-    async def update_billing(
-        workspace: UserWorkspaceDependency,
-        workspace_repository: WorkspaceRepositoryDependency,
-        billing: WorkspaceBilling,
-    ) -> WorkspaceBilling:
-        """Update a workspace billing."""
-
-        def tier(billing: WorkspaceBilling) -> SecurityTier:
-            match billing.security_tier:
-                case SecurityTierJson.Free:
-                    return SecurityTier.Free
-                case SecurityTierJson.Foundational:
-                    return SecurityTier.Foundational
-                case SecurityTierJson.HighSecurity:
-                    return SecurityTier.HighSecurity
-
-        org = await workspace_repository.update_security_tier(
-            workspace_id=workspace.id,
-            security_tier=tier(billing),
-        )
-        return WorkspaceBilling.from_model(org)
-
-    @router.put("/{workspace_id}/subscription/{subscription_id}")
-    async def assign_subscription(
-        workspace: UserWorkspaceDependency,
-        user: AuthenticatedUser,
-        subscription_repository: SubscriptionRepositoryDependency,
-        subscription_id: SubscriptionId,
-    ) -> None:
-        """Assign a subscription to a workspace."""
-        if not await subscription_repository.user_has_subscription(user.id, subscription_id):
-            raise ResourceNotFound("Subscription not found")
-
-        await subscription_repository.update_workspace(subscription_id, workspace.id)
 
     @router.get("/{workspace_id}/cf_url")
     async def get_cf_url(
