+++ tests/fixbackend/billing_information/billing_info_router_test.py
@@ -20,15 +20,19 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from fixbackend.auth.depedencies import get_current_active_verified_user
 from fixbackend.auth.models import User
 from fixbackend.config import Config, get_config
-from typing import AsyncIterator, List
+from typing import AsyncIterator, List, Optional, Sequence
 from fixbackend.app import fast_api_app
 from fixbackend.db import get_async_session
 from fixbackend.ids import BillingId, ExternalId, SecurityTier, SubscriptionId, UserId, WorkspaceId
 from fixbackend.billing_information.service import BillingEntryService, get_billing_entry_service
 from fixbackend.subscription.models import BillingEntry
+from fixbackend.subscription.subscription_repository import SubscriptionRepository, get_subscription_repository
 from fixbackend.workspaces.dependencies import get_user_workspace
 from fixbackend.workspaces.models import Workspace
 from fixcloudutils.util import utc, UTC_Date_Format
+
+from fixbackend.workspaces.repository import WorkspaceRepositoryImpl, get_workspace_repository
+from attrs import evolve
 
 external_id = ExternalId(uuid.uuid4())
 workspace_id = WorkspaceId(uuid.uuid4())
@@ -57,6 +61,8 @@ billing_entry = BillingEntry(
     reported=False,
 )
 
+sub_id = SubscriptionId(uuid.uuid4())
+
 
 class BillingEntryServiceMock(BillingEntryService):
     def __init__(self) -> None:
@@ -64,6 +70,42 @@ class BillingEntryServiceMock(BillingEntryService):
 
     async def list_billing_info(self, workspace_id: WorkspaceId) -> List[BillingEntry]:
         return [billing_entry]
+
+
+class SubscriptionRepositoryMock(SubscriptionRepository):
+    def __init__(self) -> None:
+        pass
+
+    async def user_has_subscription(self, user_id: UserId, subscription_id: SubscriptionId) -> bool:
+        return subscription_id == sub_id
+
+    async def update_workspace(self, subscription_id: SubscriptionId, workspace_id: WorkspaceId) -> None:
+        assert subscription_id == subscription_id
+        assert workspace_id == workspace_id
+        return None
+
+
+class WorkspaceRepositoryMock(WorkspaceRepositoryImpl):
+    def __init__(self) -> None:
+        pass
+
+    async def get_workspace(
+        self, workspace_id: WorkspaceId, *, session: Optional[AsyncSession] = None
+    ) -> Workspace | None:
+        return workspace
+
+    async def list_workspaces(self, owner_id: UserId) -> Sequence[Workspace]:
+        return [workspace]
+
+    async def update_workspace(self, workspace_id: WorkspaceId, name: str, generate_external_id: bool) -> Workspace:
+        if generate_external_id:
+            new_external_id = ExternalId(uuid.uuid4())
+        else:
+            new_external_id = workspace.external_id
+        return evolve(workspace, name=name, external_id=new_external_id)
+
+    async def update_security_tier(self, workspace_id: WorkspaceId, security_tier: SecurityTier) -> Workspace:
+        return evolve(workspace, security_tier=security_tier)
 
 
 @pytest.fixture
@@ -75,6 +117,8 @@ async def client(session: AsyncSession, default_config: Config) -> AsyncIterator
     app.dependency_overrides[get_user_workspace] = lambda: workspace
     app.dependency_overrides[get_current_active_verified_user] = lambda: user
     app.dependency_overrides[get_billing_entry_service] = lambda: BillingEntryServiceMock()
+    app.dependency_overrides[get_subscription_repository] = lambda: SubscriptionRepositoryMock()
+    app.dependency_overrides[get_workspace_repository] = lambda: WorkspaceRepositoryMock()
 
     async with AsyncClient(app=app, base_url="http://test") as ac:
         yield ac
@@ -93,3 +137,28 @@ async def test_list_billing_entries(client: AsyncClient) -> None:
     assert json_billing_entry["period_start"] == billing_entry.period_start.strftime(UTC_Date_Format)
     assert json_billing_entry["period_end"] == billing_entry.period_end.strftime(UTC_Date_Format)
     assert json_billing_entry["nr_of_accounts_charged"] == billing_entry.nr_of_accounts_charged
+
+
+@pytest.mark.asyncio
+async def test_billing(client: AsyncClient) -> None:
+    response = await client.get(f"/api/workspaces/{workspace_id}/billing")
+    assert response.json().get("payment_method") == "aws_marketplace"
+    assert response.json().get("security_tier") == "free"
+
+    update_response = await client.put(
+        f"/api/workspaces/{workspace_id}/billing",
+        json={"payment_method": "aws_marketplace", "security_tier": "foundational"},
+    )
+
+    assert update_response.json().get("payment_method") == "aws_marketplace"
+    assert update_response.json().get("security_tier") == "foundational"
+
+
+@pytest.mark.asyncio
+async def test_update_subscription(client: AsyncClient) -> None:
+    response = await client.put(f"/api/workspaces/{workspace_id}/subscription/{sub_id}")
+    assert response.status_code == 200
+
+    unknown_subscription = SubscriptionId(uuid.uuid4())
+    response = await client.put(f"/api/workspaces/{workspace_id}/subscription/{unknown_subscription}")
+    assert response.status_code == 404
